Create FUNDING.yml
We are UK based company that has experienced a number of difficulties when charity money is being sent to warn torn countries to help refugees. The time it takes to send  financial assistance to  third world countries in an emergency situation is extremely slow. We have developed a coin design and a name for the coin . What we require is a developer who can help us develop the crypto based system that can be utilised on a rapid payment system.